Add Google Closure Compiler JS version
Keeping for now the java version
