+++ lib/compress.js
@@ -14,6 +14,7 @@ var fs = require('fs');
 var mkdirp = require('mkdirp');
 var cleanCSS = require('./compressors/clean-css');
 var csso = require('./compressors/csso');
+var gccJava = require('./compressors/gcc-java');
 var gcc = require('./compressors/gcc');
 var noCompress = require('./compressors/no-compress');
 var sqwish = require('./compressors/sqwish');
@@ -35,11 +36,12 @@ var compressorsMap = {
   'yui-js': function(settings, data) {
     return yui('js', settings, data);
   },
-  'gcc': function(settings, data) {
-    return gcc(settings, data);
+  'gcc': gcc,
+  'gcc-java': function(settings, data) {
+    return gccJava(settings, data, false);
   },
   'gcc-legacy': function(settings, data) {
-    return gcc(settings, data, true);
+    return gccJava(settings, data, true);
   },
   'uglifyjs': uglifyjs,
   'sqwish': sqwish,
