+++ lib/compressors/gcc.js
@@ -10,57 +10,27 @@
  * Module dependencies.
  */
 
-var path = require('path');
+var Promise = require('bluebird');
+var gcc = require('google-closure-compiler-js');
 var utils = require('../utils');
-var execCompressor = require('../runner');
 
 /**
- * Module variables.
- */
-var binGcc = path.normalize(__dirname + '/../binaries/google_closure_compiler-v20151216.jar');
-var binGccLegacy = path.normalize(__dirname + '/../binaries/google_closure_compiler-v20131014-legacy-java-1.6.jar');
-
-/**
- * Expose `compressGCC()`.
+ * Expose `compressGCCJS()`.
  */
 
-module.exports = compressGCC;
+module.exports = compressGCCJS;
 
 /**
  * Run Google Closure Compiler.
  *
  * @param {Object} settings
  * @param {String} content
- * @param {Boolean} legacy
  */
 
-function compressGCC(settings, content, legacy) {
-  return execCompressor(gccCommand(settings.options, legacy), content, settings)
-    .then(function(contentMinified) {
-      return utils.writeFile(settings.output, contentMinified);
-    })
-    .catch(handleErrors);
-}
-
-function handleErrors(err) {
-  if (err.message.indexOf('UnsupportedClassVersionError') > -1) {
-    throw new Error('Latest Google Closure Compiler requires Java >= 1.7, please update Java or use gcc-legacy');
-  }
-  throw err;
-}
-
-/**
- * Google Closure Compiler command line.
- */
-
-function gccCommand(options, legacy) {
-  return [
-    '-server',
-    '-XX:+TieredCompilation',
-    '-jar',
-    '-Xss2048k',
-    (legacy) ? binGccLegacy : binGcc,
-    '--warning_level',
-    'QUIET'
-  ].concat(utils.buildArgs(options));
+function compressGCCJS(settings, content) {
+  return new Promise(function(resolve) {
+    var contentMinified = gcc.compile({jsCode: [{src: content}]}); // TODO add settings
+    utils.writeFile(settings.output, contentMinified.compiledCode);
+    resolve(contentMinified.compiledCode);
+  });
 }
